refactor(pairings): relocate summary metric
